refactor(Navbar): simplify menu and authentication handling; add new components for user actions
